[SC-332] improve SSM session experience
My initial idea to alleviate confusion around the one time setup of files
needed to run the aws session manager is to setup a GH repo with all
needed files ready to go which should encompes all of Tess Thyer suggestions
from issue SC-183.   This is just an initial step to make things easier for
users then we can gather some feedback to see if we need further automation.